src & dependencies for @fireloop/<module>
