+++ README.md
@@ -1,5 +1,8 @@
 ![FireLoop.io](https://storage.googleapis.com/mean-expert-images/fireloop-logo.png)
 
+
+[![Deploy](https://www.herokucdn.com/deploy/button.svg)](https://heroku.com/deploy?template=https://github.com/mean-expert-official/fireloop-starter/tree/master)
+
 ##Presenting FireLoop
 
 [FireLoop] is a [NodeJS] Real-Time Platform that allows you to build modern and complex applications by seamlessly integrating the amazing technologies from the MEAN Stack; Like the IBM's [LoopBack], Google's [Angular 2], Telerik's [NativeScript 2] and [Ionic 2] Frameworks.
@@ -21,10 +24,6 @@
 - [Ionic 2] Integration.
 - [SDK Builder] Integration.
 - [Real-Time] Integration.
-
-##Deploy in heroku
-
-[![Deploy](https://www.herokucdn.com/deploy/button.svg)](https://heroku.com/deploy?template=https://github.com/mean-expert-official/fireloop-starter/tree/master)
 
 ##Installing FireLoop CLI
