--- core/decorators/boot-script/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@mean-expert/boot-script",
-  "version": "1.0.0",
-  "description": "FireLoop BootScript Decorator",
-  "main": "index.js",
-  "types": "index.d.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "files": [
-    "index.js",
-    "index.js.map",
-    "index.d.ts",
-    "package.json"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mean-expert-official/boot-script.git"
-  },
-  "keywords": [
-    "FireLoop"
-  ],
-  "author": "Jonathan Casarrubias",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/mean-expert-official/boot-script/issues"
-  },
-  "homepage": "https://github.com/mean-expert-official/boot-script#readme",
-  "dependencies": {},
-  "devDependencies": {}
-}
