--- core/decorators/inject/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@mean-expert/inject",
-  "version": "1.0.0",
-  "description": "FireLoop Inject Decorator",
-  "main": "index.js",
-  "types": "index.d.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "files": [
-    "index.js",
-    "index.js.map",
-    "index.d.ts",
-    "package.json"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mean-expert-official/boot-script.git"
-  },
-  "keywords": [
-    "index.js",
-    "index.js.map",
-    "index.d.ts",
-    "package.json"
-  ],
-  "author": "Jonathan Casarrubias",
-  "license": "MIT",
-  
-  "bugs": {
-    "url": "https://github.com/mean-expert-official/boot-script/issues"
-  },
-  "homepage": "https://github.com/mean-expert-official/boot-script#readme",
-  "dependencies": {},
-  "devDependencies": {}
-}
