--- core/decorators/inject/tsconfig.json
@@ -1,14 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "target": "es5",
-        "moduleResolution": "node",
-        "noImplicitAny": true,
-        "removeComments": true,
-        "preserveConstEnums": true,
-        "sourceMap": true
-    },
-    "files": [
-        "index.ts"
-    ]
-}
