--- core/decorators/model/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@mean-expert/model",
-  "version": "1.0.9",
-  "description": "FireLoop Model Decorator",
-  "main": "index.js",
-  "types": "index.d.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "files": [
-    "index.js",
-    "index.js.map",
-    "index.d.ts",
-    "package.json"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mean-expert-official/model.git"
-  },
-  "keywords": [
-    "FireLoop"
-  ],
-  "author": "Jonathan Casarrubias",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/mean-expert-official/model/issues"
-  },
-  "homepage": "https://github.com/mean-expert-official/model#readme",
-  "dependencies": {
-    "@fireloop/lib": "lastest"
-  }
-}
