+++ core/index.ts
@@ -1,3 +1,1 @@
-export * from './decorators/model';
-export * from './decorators/inject';
-export * from './decorators/boot-script';
+export * from './src/decorators';
