+++ core/src/decorators/boot-script/index.d.ts
