+++ core/src/decorators/boot-script/index.ts
