+++ core/src/decorators/index.ts
@@ -1,0 +1,3 @@
+export * from './model';
+export * from './inject';
+export * from './boot-script';
