+++ core/src/decorators/inject/dependencies.ts
@@ -1,4 +1,4 @@
-import {Injectable}       from    './injectable';
+import { Injectable } from './injectable';
 export interface Dependencies {
     providers: Array<Injectable>;
 }
