+++ core/src/decorators/inject/index.d.ts
