+++ core/src/decorators/model/index.d.ts
