+++ lib/.npmignore
@@ -1,0 +1,3 @@
+node_modules
+src
+tests
