--- lib/model-register/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@mean-expert/model-register",
-  "version": "1.0.5",
-  "description": "FireLoop Model Register Module",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "files": [
-    "index.js",
-    "index.js.map",
-    "package.json"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mean-expert-official/model-register.git"
-  },
-  "keywords": [
-    "FireLoop"
-  ],
-  "author": "Jonathan Casarrubias",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/mean-expert-official/model-register/issues"
-  },
-  "homepage": "https://github.com/mean-expert-official/model-register#readme"
-}
