--- lib/model-register/tsconfig.json
@@ -1,12 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "noImplicitAny": true,
-        "removeComments": true,
-        "preserveConstEnums": true,
-        "sourceMap": true
-    },
-    "files": [
-        "index.ts"
-    ]
-}
