+++ lib/src/model-register/index.ts
