+++ lib/tsconfig.json
@@ -1,0 +1,26 @@
+{
+  "compilerOptions": {
+    "declaration": false,
+    "emitDecoratorMetadata": true,
+    "experimentalDecorators": true,
+    "noImplicitAny": true,
+    "mapRoot": "src",
+    "module": "commonjs",
+    "outDir": "./",
+    "sourceMap": true,
+    "target": "es5",
+    "typeRoots": [
+      "../node_modules/@types"
+    ],
+    "types": [
+      "chalk",
+      "ejs",
+      "node"
+    ]
+  },
+  "exclude": [
+    "node_modules",
+    "tests",
+    "**/*.spec.ts"
+  ]
+}
